backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "atlas-agentic-rag-backend"
version = "0.1.0"
description = "Atlas Agentic RAG Copilot backend skeleton"
readme = "../README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.16.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "langgraph>=0.2.0",
  "orjson>=3.10.0",
  "pgvector>=0.3.6",
  "pydantic-settings>=2.8.0",
  "pypdf>=5.4.0",
  "psycopg[binary]>=3.2.0",
  "python-docx>=1.1.0",
  "python-multipart>=0.0.9",
  "python-pptx>=1.0.2",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "pytest-cov>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
